requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements/requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
    #   openai
blinker==1.9.0
    # via flask
cachetools==5.5.2
    # via google-auth
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via flask
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
distro==1.9.0
    # via openai
exceptiongroup==1.3.0
    # via anyio
flask==3.1.0
    # via -r requirements/requirements.in
fonttools==4.56.0
    # via matplotlib
google-auth==2.40.3
    # via google-genai
google-genai==1.20.0
    # via -r requirements/requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   google-genai
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jiter==0.10.0
    # via openai
joblib==1.4.2
    # via
    #   mlxtend
    #   scikit-learn
kiwisolver==1.4.8
    # via matplotlib
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.10.1
    # via mlxtend
mlxtend==0.23.4
    # via -r requirements/requirements.in
networkx==3.4.2
    # via
    #   -r requirements/requirements.in
    #   python-louvain
numpy==2.2.3
    # via
    #   -r requirements/requirements.in
    #   contourpy
    #   matplotlib
    #   mlxtend
    #   pandas
    #   python-louvain
    #   scikit-learn
    #   scipy
openai==1.86.0
    # via -r requirements/requirements.in
packaging==24.2
    # via matplotlib
pandas==2.2.3
    # via
    #   -r requirements/requirements.in
    #   mlxtend
pillow==11.1.0
    # via matplotlib
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.6
    # via
    #   google-genai
    #   openai
pydantic-core==2.33.2
    # via pydantic
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.1.0
    # via -r requirements/requirements.in
python-louvain==0.16
    # via -r requirements/requirements.in
pytz==2025.1
    # via pandas
requests==2.32.4
    # via google-genai
rsa==4.9.1
    # via google-auth
scikit-learn==1.6.1
    # via mlxtend
scipy==1.15.2
    # via
    #   mlxtend
    #   scikit-learn
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
threadpoolctl==3.6.0
    # via scikit-learn
tqdm==4.67.1
    # via openai
typing-extensions==4.14.0
    # via
    #   anyio
    #   exceptiongroup
    #   google-genai
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.1
    # via pandas
urllib3==2.4.0
    # via requests
websockets==15.0.1
    # via google-genai
werkzeug==3.1.3
    # via flask
